Restructure package to align holland.lib.common with import path
The current structure wasn't compatible with some packaging tools when
I attempted to incorporate a pyproject.toml across the source tree. This
corrects the structure so that it is consistent with other holland modules.
