+++ plugins/holland.backup.mongodump/holland/backup/mongodump.py
@@ -10,8 +10,8 @@ from io import open  # pylint: disable=redefined-builtin
 from pymongo import MongoClient
 
 from holland.core.backup import BackupError
-from holland.lib.compression import COMPRESSION_CONFIG_STRING, open_stream
-from holland.lib.which import which
+from holland.lib.common.compression import COMPRESSION_CONFIG_STRING, open_stream
+from holland.lib.common.which import which
 
 LOG = logging.getLogger(__name__)
