+++ plugins/holland.backup.mysqldump/holland/backup/mysqldump/base.py
@@ -6,7 +6,7 @@ import logging
 
 from holland.backup.mysqldump.command import ALL_DATABASES
 from holland.core.backup import BackupError
-from holland.lib.safefilename import encode
+from holland.lib.common.safefilename import encode
 
 LOG = logging.getLogger(__name__)
