+++ plugins/holland.backup.tar/holland/backup/tar.py
@@ -8,7 +8,7 @@ from subprocess import Popen, list2cmdline
 from tempfile import TemporaryFile
 
 from holland.core.backup import BackupError
-from holland.lib.compression import COMPRESSION_CONFIG_STRING, open_stream
+from holland.lib.common.compression import COMPRESSION_CONFIG_STRING, open_stream
 
 LOG = logging.getLogger(__name__)
