+++ plugins/holland.backup.xtrabackup/holland/backup/xtrabackup/util.py
@@ -16,7 +16,7 @@ from string import Template
 from subprocess import PIPE, STDOUT, Popen, list2cmdline
 
 from holland.core.backup import BackupError
-from holland.lib.which import which
+from holland.lib.common.which import which
 
 LOG = logging.getLogger(__name__)
