+++ plugins/holland.lib.common/holland/lib/__init__.py
@@ -1,5 +1,5 @@
 """
-Common plugin utils
+Holland Common plugins
 """
 
 __import__("pkg_resources").declare_namespace(__name__)
